(PUP-6797) Update SMF tests for nonexistent service
Prior to this commit, the puppet tests would check for puppet output indicating
a previously nonexistent smf service by checking that it had transitioned from
'stopped' to 'running'. However, the fact that puppet would output `ensure =>
stopped` for non-existent services was in error, addressed as PUP-6797 - it
should have been `ensure => absent` as it is with other platforms. This commit
updates the various acceptance tests to instead grep for the now-correct
transition, which is the definition of ensure as running.

Signed-off-by: Moses Mendoza <moses@puppet.com>